[trunk] - Update NuGet.exe - Updated version information - Applied patch from kjermey [issue:139, ReSharper suggestions]
Issue: 139


git-svn-id: http://nettopologysuite.googlecode.com/svn/trunk@999 53ca42c3-9f19-0410-8ccb-b990a9bb5db6

+++ NetTopologySuite/Index/Chain/MonotoneChainBuilder.cs
@@ -27,7 +27,7 @@ namespace NetTopologySuite.Index.Chain
         {
             int[] array = new int[list.Count];
             for (int i = 0; i < array.Length; i++)            
-                array[i] = (int)list[i];            
+                array[i] = list[i];            
             return array;
         }
